Whole Project: Added mocha unit tests support, npm run test
