+++ package.json
@@ -3,7 +3,8 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "nodemon ./bin/www"
+    "start": "nodemon ./bin/www",
+    "test": "mocha -r ts-node/register tests/**/*.ts"
   },
   "dependencies": {
     "chai": "^4.2.0",
@@ -19,6 +20,8 @@
     "watchify": "^4.0.0"
   },
   "devDependencies": {
+    "@types/chai": "^4.3.1",
+    "@types/mocha": "^9.1.1",
     "typescript": "^4.7.3"
   }
 }
