+++ tests/src/testFile.ts
@@ -1,0 +1,8 @@
+import {expect} from 'chai';
+
+describe('hello', ()=>{
+
+    it('should...', ()=>{
+        expect(1).to.equal(1);
+    });
+});
