+++ tests/temptester.js
@@ -1,0 +1,7 @@
+
+
+ const typeSignatures = require('../src/types/typeSignatures').typeSignatures;
+ const expect = require('chai').expect; 
+ 
+ 
+let ts = typeSignatures.signature(['A']) ;
