+++ tests/typeRegister.js
