+++ tests/typeSignatures/main.js
